Refactor: Move login loading overlay logic from App.tsx to SignInScreen
- Move GlobalLoadingOverlay from App.tsx to SignInScreen for better separation of concerns
- Move getLoadingMessage function to SignInScreen where it belongs
- Simplify App.tsx to only render TabNavigation
- Follow single responsibility principle - login UI logic stays in login screen

This improves code organization by keeping login-specific logic contained
within the SignInScreen component rather than at the app level.
